Your program should:
Read an XML file containing news stories and headlines.
Extract the headers and the text.
Tokenize the texts.
For each story, find the most frequent tokens that appear in it.
Print the headline of each news story and the five most frequent tokens in descending order. Take a look at a sample below. Also, display the titles and keywords in the same order they are presented in the file.