Your program should:
Read an XML file containing news stories and headlines.
Extract the headers and the text.
Tokenize each text.
Lemmatize each word in the story.
Get rid of punctuation and stopwords with the help of NLTK.
For each news story, find the most frequent tokens.
Print each story's headline and five most frequent tokens in descending order. Take a look at the sample below. Display the titles and keywords in the same order they are presented in the file.