This tricky sentence is fine for us, but may be confusing for the tagger:
sent = ['The', 'horse', 'raced', 'past', 'the', 'barn', 'fell', 'down', '.']

If we tag the sentence above, the word 'raced' will have the tag 'VBD'. This tag is not correct, but we can make the sentence easier for the tagger:

sent = ['The', 'horse', 'that', 'was', 'raced', 'past', 'the', 'barn', 'fell', 'down', '.']

In such case, the tag of the word 'raced' will be correct. Write a code to tag this sentence with the help of NLTK and print the result for the word 'raced' now.