In this stage, your program should:
Read an XML-file containing news stories and headlines.
Extract the headers and the text.
Tokenize each text.
Lemmatize each word in the story.
Get rid of punctuation and the stopwords provided by NLTK.
For each news story, find the most frequent nouns.
Print each story's headline and the five most frequent nouns in descending order. Take a look at the sample below. Display the titles and keywords in the same order they are presented in the file.