Suppose, we have a movie database and we want to extract the day of release. The entry for the date is stored like this:
Day of release: 4 July 2019
Your task is to finish the function that extracts the date from this string. The function is called get_release_date. It takes the release day entry in the described format (a string) and returns the corresponding datetime object.

Note that you do NOT need to call the function, process the input, or print anything. The examples are provided for you to test your code.