In the topic, we didn't cover all the format codes. Check out the official documentation to find the code used to format the week number of the year (assuming that Sunday is the first day of the week).
Write a function get_week_number that takes a datetime object and returns the string "Week number: {num}." where num represents the week number.

Note that you do NOT need to read the input, call the function, or print the result. Input and output example is provided for you to test your code.